add additional buffering modes @ lab 4
Each file now is able to hold exclusive access to
buffer for entire word, or even a line.

Buffering mode also has human-readable toString()
message,so it can be used in CLI options etc.

For eventual future-proofing, new TwoWayMonitor is
used, so that waiting for buffer read and buffer
write can happen independently.

Additional tests were written, and existing ones
updated, as noGui-mode is needed for when toolkit
is not initialized.